requirements.txt
requests==2.30.0
Appium-Python-Client==4.2.1
allure-python-commons==2.13.5
allure-pytest==2.13.5
pytest-asyncio==0.24.0
pymysql==1.1.1
webdrivermanager_cn==2.0.1
xlrd== 2.0.1
configparser==5.2.0
setuptools==75.2.0
Cython==3.0.11
pytest==8.3.3
selenium==4.18.1
playwright==1.48.0